pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "p2d"
description = "Simple python script converting polygon package to domjudge(kattis) package"
authors = [{ name = "cubercsl", email = "hi@cubercsl.site" }]
maintainers = [
    { name = "cubercsl", email = "hi@cubercsl.site" },
    { name = "Dup4", email = "hi@dup4.com" },
]
readme = "README.md"
license = "MIT"
keywords = ["Polygon", "DOMjudge", "Kattis", "Codeforces"]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "deepmerge>=2.0",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "tomli>=2.0.1;python_version<'3.11'",
    "typer-slim[standard]>=0.12.5",
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/cn-xcpc-tools/Polygon2Domjudge"
documentation = "https://github.com/cn-xcpc-tools/Polygon2Domjudge"
repository = "https://github.com/cn-xcpc-tools/Polygon2Domjudge"

[project.scripts]
p2d = "p2d.cli:app"
p2d-contest = "p2d.contest:app"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "types-pyyaml>=6.0.12.20240917",
    "mypy>=1.11.2",
    "pytest-cov>=5.0.0",
]

[tool.ruff]
line-length = 127

[tool.pytest.ini_options]
addopts = [
    "--cov=p2d",
    "--cov-report",
    "term-missing",
    "--cov-report",
    "xml",
    "--cov-fail-under=80",
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
    "raise ProcessError",
]

[tool.hatch.build.hooks.vcs]
version-file = "p2d/_version.py"

[tool.hatch.build.targets.wheel]
exclude = ["p2d/testlib/**/*", "!p2d/testlib/testlib.h", "!p2d/testlib/LICENSE"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
git_describe_command = [
    "git",
    "describe",
    "--dirty",
    "--tags",
    "--long",
    "--first-parent",
]

[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true
